Refactor article deletion to remove DeleteArticleUseCase

+++ src/main/kotlin/io/github/gunkim/realworld/domain/article/service/DeleteArticleService.kt
@@ -1,16 +1,24 @@
 package io.github.gunkim.realworld.domain.article.service
 
-import io.github.gunkim.realworld.domain.article.model.Article
+import io.github.gunkim.realworld.domain.article.ArticleFindable
+import io.github.gunkim.realworld.domain.article.model.Slug
 import io.github.gunkim.realworld.domain.article.repository.ArticleRepository
+import io.github.gunkim.realworld.domain.user.model.UserId
 import org.springframework.stereotype.Service
 
 @Service
 class DeleteArticleService(
-    private val articleRepository: ArticleRepository,
-) {
+    override val articleRepository: ArticleRepository,
+    private val articleOwnershipService: ArticleOwnershipService,
+) : ArticleFindable {
     fun deleteArticle(
-        article: Article,
+        slug: Slug,
+        deleterId: UserId,
     ) {
+        val article = findBySlug(slug)
+
+        articleOwnershipService.validateOwnership(article, deleterId)
+
         articleRepository.delete(article)
     }
 }
